Filtered invoices using the orgId if provided, else fetching using userId on invoices with a null orgId

+++ app/actions.ts
@@ -3,7 +3,7 @@
 import { db } from "@/db";
 import { Customers, Invoices } from "@/db/schema";
 import { auth } from "@clerk/nextjs/server";
-import { and, eq } from "drizzle-orm";
+import { and, eq, isNull } from "drizzle-orm";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
 
@@ -46,25 +46,33 @@ export async function createAction(formData: FormData) {
 
 export async function updateStatusAction(formData: FormData) {
     console.log("Entered the updateStatusAction function")
-    const { userId } = await auth()
+    const { userId, orgId } = await auth()
     if(!userId) return;
 
     const id = formData.get('id') as string;
     const status = formData.get('status') as string;
-
-    await db.update(Invoices).set({status}).where(and(eq(Invoices.id, parseInt(id)),eq(Invoices.userId, userId)))
+    if(orgId){
+        await db.update(Invoices).set({status}).where(and(eq(Invoices.id, parseInt(id)),eq(Invoices.userId, userId), eq(Invoices.organizationId, orgId)))
+    } else {
+        await db.update(Invoices).set({status}).where(and(eq(Invoices.id, parseInt(id)),eq(Invoices.userId, userId), isNull(Invoices.organizationId)))
+    }
+    
 
 
     revalidatePath(`/invoices/${id}`, 'page' )
 }
 
 export async function deleteInvoiceAction(formData: FormData){
-    const { userId } = await auth()
+    const { userId, orgId } = await auth()
     if(!userId) return;
 
     const id = formData.get('id') as string;
 
-    await db.delete(Invoices).where(and(eq(Invoices.id, parseInt(id)),eq(Invoices.userId, userId)))
+    if(orgId) {
+        await db.delete(Invoices).where(and(eq(Invoices.id, parseInt(id)),eq(Invoices.userId, userId), eq(Invoices.organizationId, orgId)))
+    } else {   
+        await db.delete(Invoices).where(and(eq(Invoices.id, parseInt(id)),eq(Invoices.userId, userId), isNull(Invoices.organizationId)))
+    }
 
 
     redirect("/dashboard" )
